Rename profiles/profilesSlice to profile/profileSlice
